requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
PyGithub==1.59.1
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
